+++ zfs/replicate/receive/__init__.py
@@ -1,0 +1,3 @@
+"""ZFS Receive Operations."""
+
+from .type import Options as Options  # noqa: F401 # pylint: disable=C0414