+++ zfs_test/replicate_test/receive_test/__init__.py
@@ -1,0 +1,1 @@
+"""Test receive functionality."""